requirements.txt
Flask>=3.1.1
flask-cors>=6.0.1
requests>=2.32.3
beautifulsoup4>=4.13.3
openai>=1.97.0
python-dotenv>=1.1.0
sentence-transformers>=5.0.0
keybert>=0.9.0
spacy>=3.8.7
numpy>=2.2.6
PyMuPDF>=1.26.3
python-docx>=1.2.0
gunicorn>=23.0.0
google-generativeai>=0.8.5
torch>=2.7.1